requirements.txt
APScheduler==3.6.3
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
python-telegram-bot==13.5
pytz==2024.1
six==1.16.0
tornado==6.4.1
tzlocal==5.2
Werkzeug==3.0.3
